requirements.txt
# Core Framework
agency-swarm>=0.1.0
fastapi>=0.104.1
pydantic>=2.5.2

# Database and Storage
duckdb>=0.9.2
redis>=5.0.1
pymilvus>=2.3.4

# AI and Research
anthropic>=0.7.7
tavily-python>=0.2.6

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.1
asyncio>=3.4.3

# Testing and Development
pytest>=7.4.3
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.4.14 